% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy2c.R
\name{copy2c}
\alias{copy2c}
\title{Copy data.frame R ke dalam Clipboard}
\usage{
copy2c(df, dec = ",", rownames = F, colnames = F, ...)
}
\arguments{
\item{df}{DataFrame yang ingin di copy}

\item{dec}{format desimal, default dengan koma}

\item{rownames}{apakah nama baris ikut dicopy?}

\item{colnames}{apakah nama kolom ikut dicopy?}

\item{...}{parameter untuk \code{write.table}}
}
\value{
tidak ada (silahkan coba paste di spreadshet atau lainnya)
}
\description{
Copy data.frame R ke dalam Clipboard
}
\examples{
copy2c(iris)
}
