% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\docType{package}
\name{dm}
\alias{dm}
\title{dm: A package for data manipulation}
\description{
The dm package provides three categories of important functions:
str2vec, read_img and many more
}
\section{Foo functions}{

The foo functions ...
}

