% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_conflict_prefer.R
\name{my_conflict_prefer}
\alias{my_conflict_prefer}
\title{Setting default conflict ke preferensi saya}
\usage{
my_conflict_prefer(quiet = TRUE)
}
\arguments{
\item{quiet}{logical}
}
\value{
none
}
\description{
Setting default conflict ke preferensi saya
}
\examples{
my_conflict_prefer()
}
