% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_img.R
\name{read_img}
\alias{read_img}
\title{Membaca data dari gambar}
\usage{
read_img(path = "clipboard", bahasa = "eng", to_df = FALSE, ...)
}
\arguments{
\item{path}{alamat path dari file, default akan membaca dari clipboard}

\item{bahasa}{bahasa yang digunakan 'eng', 'ind', 'numbers'}

\item{to_df}{logical, jika TRUE maka output akan langsung berupa data.frame}

\item{...}{input untuk fungsi write.table()}
}
\value{
hasil berupa character atau data.frame
}
\description{
Membaca data dari gambar
}
\examples{
read_img() # membaca dari clipboard
read_img("D:/img.jpg", "numbers")
}
