% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_string.R
\name{read_string}
\alias{read_string}
\title{Membaca data.frame dari string}
\usage{
read_string(s, col_names = NULL, header = FALSE, tranpose = FALSE, ...)
}
\arguments{
\item{s}{string}

\item{col_names}{nama kolom yang diigninkan, default adalah V1, V2 ...}

\item{header}{logical. apakah data.frame menggandung header}

\item{tranpose}{logical. apakah perlu tranpose}

\item{...}{parameter untuk \code{read.table}}
}
\value{
data.frame
}
\description{
Membaca data.frame dari string
}
\examples{
s1 <-
  "A 112 20
 YY 18 10
 XX 15 15
 DD 25 28"

read_string(s1)
read_string(s1, header = TRUE)
read_string(s1, tranpose = TRUE, header = FALSE)
read_string(s1, tranpose = TRUE, header = TRUE, col_names = LETTERS[1:4])
read_string(s1, tranpose = TRUE, col_names = letters[1:4])
}
