% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separator_convert.R
\name{separator_convert}
\alias{separator_convert}
\title{Merubah Separator dari beberapa kolom tipe string secara sekaligus}
\usage{
separator_convert(dat, col, sep = ",", to_sep = "\\\\.", to_numeric = T)
}
\arguments{
\item{dat}{data.frame atau tibble}

\item{col}{kolom yang ingin dirubah}

\item{sep}{pemisah asal, default adalah koma}

\item{to_sep}{pemisah tujuan, default adalah titik}

\item{to_numeric}{ubah kolom tersebut menjadi numeric ?}
}
\value{
data.frame
}
\description{
Merubah Separator dari beberapa kolom tipe string secara sekaligus
}
\examples{

# Input String
s2 <- "Melalui Transaksi Penjualan	9	4,6	28	14,4	16	8,2
Tidak Melalui Transaksi Penjualan	24	12,4	85	43,8	32	16,5
Melalui Transaksi Penjualan	14	14,1	24	24,2	17	17,2
Tidak Melalui Transaksi Penjualan	10	10,1	18	18,2	16	16,2"

dat <- read_pattern(s2, pos_non_angka = 1, pos_angka = 2:7)
dat <- separator_convert(dat, v2:v7)
dat
}
