% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2vec.R
\name{str2vec}
\alias{str2vec}
\title{Merubah string menjadi vektor}
\usage{
str2vec(str, pemisah = " ")
}
\arguments{
\item{str}{character}

\item{pemisah}{character}
}
\value{
vector
}
\description{
Merubah string menjadi vektor
}
\examples{
str2vec('1.2 3.4 5 9 100')
str2vec('1.2, 3.4, 5, 9, 100', pemisah = '\\\\,')
}
